+++ ginga/cmap.py
@@ -4,17 +4,11 @@
 # This is open-source software licensed under a BSD license.
 # Please see the file LICENSE.txt for details.
 #
-import warnings
-
 import numpy as np
-from astropy.utils import minversion
 
 __all__ = ['ColorMap', 'add_cmap', 'get_cmap', 'has_cmap', 'get_names',
            'matplotlib_to_ginga_cmap', 'add_matplotlib_cmap',
            'add_matplotlib_cmaps']
-
-MPL_LT_3_4 = not minversion('matplotlib', '3.4')
-
 
 # Some built in colormaps
 
@@ -13309,12 +13303,7 @@ def add_matplotlib_cmap(cm, name=None):
 def add_matplotlib_cmaps(fail_on_import_error=True):
     """Add all matplotlib colormaps."""
     try:
-        import matplotlib.pyplot as plt
-        from matplotlib import cm as _cm
-        if MPL_LT_3_4:
-            from matplotlib.cbook import mplDeprecation as MatplotlibDeprecationWarning
-        else:
-            from matplotlib import MatplotlibDeprecationWarning
+        from matplotlib import colormaps as _mpl_cm
     except ImportError:
         if fail_on_import_error:
             raise
@@ -13322,15 +13311,12 @@ def add_matplotlib_cmaps(fail_on_import_error=True):
         return
 
     # NOTE: Update if matplotlib has new public API for this.
-    for name in plt.colormaps():
+    for name in _mpl_cm:
         if not isinstance(name, str):
             continue
         try:
-            # Do not load deprecated colormaps
-            with warnings.catch_warnings():
-                warnings.simplefilter('error', MatplotlibDeprecationWarning)
-                cm = _cm.get_cmap(name)
-                add_matplotlib_cmap(cm, name=name)
+            cm = _mpl_cm[name]
+            add_matplotlib_cmap(cm, name=name)
         except Exception as e:
             if fail_on_import_error:
                 print(f"Error adding colormap '{name}': {e!r}")