+++ doc/WhatsNew.rst
@@ -21,6 +21,7 @@ Ver 5.0.0 (unreleased)
 - Fix for a logger annoyance message when mousing over far edge of image
 - Updates for deprecations in numpy 2.0
 - Fix for missing menubar on some versions of Qt and Mac OS X
+- Fix for importing mpl colormaps with recent versions of matplotlib
 
 Ver 4.1.0 (2022-06-30)
 ======================